requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
boto3==1.26.77
botocore==1.29.77
Django==4.1.7
django-bootstrap4==22.3
django-environ==0.9.0
django-storages==1.13.2
gunicorn==20.1.0
jmespath==1.0.1
Pillow==9.4.0
psycopg2-binary==2.9.5
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
urllib3==1.26.14